requirements.txt
numpy
nltk
spacy
requests
matplotlib
seaborn
networkx
boto3
python-dotenv
pandas
ipykernel
ipywidgets
jupyter
tqdm